refactor: move macos rust logic to macos package directory
