+++ src-tauri/src/display.rs
@@ -6,7 +6,7 @@ use tauri_plugin_theme::ThemePlugin;
 use tauri_plugin_autostart::MacosLauncher;
 
 use crate::app::AppState;
-use crate::generators::macos::set_transparent_titlebar;
+use crate::macos::set_transparent_titlebar;
 use crate::helpers::fs::load_settings;
 use crate::state::Settings;
 use crate::ui::tray::{MAIN_WINDOW_LABEL, setup_tray};
@@ -70,7 +70,7 @@ pub fn create_app<R: Runtime>(app: AppState, builder: tauri::Builder<R>) -> taur
                     state.emit_disks(&win);
                     state.emit_processes(&win);
                     state.emit_batteries(&win);
-                    tokio::time::sleep(Duration::from_secs(1)).await;
+                    tokio::time::sleep(Duration::from_secs(10)).await;
                 }
             });
