--- src-tauri/src/generators/mod.rs
@@ -1,1 +1,0 @@
-pub mod macos;
