+++ src-tauri/src/lib.rs
@@ -4,7 +4,7 @@ pub mod utils;
 pub mod metrics;
 pub mod app;
 pub mod ui;
-pub mod generators;
+pub mod macos;
 pub mod commands;
 pub mod state;
 pub mod helpers;
